requirements.txt
flask
google-generativeai
supabase
python-dotenv
